+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.3 Under development
 
 - Bug #23: Add `Super-Linter` badge to `README.md` for improved visibility on code quality checks (@terabytesoftw)
+- Bug #24: Replace `MockerFunctions` with `MockerState` class for `microtime()` handling and improve test isolation (@terabytesoftw)
 
 ## 0.1.2 January 27, 2026